% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IC.R
\name{IC}
\alias{IC}
\title{Information Criterion for selecting the tuning parameter.}
\usage{
IC(obj, criterion = c("BIC", "AIC", "GCV", "AICc", "EBIC"))
}
\arguments{
\item{obj}{A fitted LassoGEE object.}

\item{criterion}{The criterion by which to select the regularization parameter.
One of "AIC", "BIC", "GCV", "AICc", or "EBIC"; default is "BIC".}
}
\value{
\item{IC}{The calculated model selection criteria}
}
\description{
Information Criterion for a fitted LassoGEE object with
 the AIC, BIC, or GCV criteria.
}
\references{
Gao, X., and Yi, G. Y. (2013). Simultaneous model selection and estimation
for mean and association structures with clustered binary data. Stat, 2(1), 102-118.
}
