% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEE_L1.R
\name{LassoGEE}
\alias{LassoGEE}
\title{Function to fit penalized GEE by I-CGD algorithm.}
\usage{
LassoGEE(
  X,
  y,
  id,
  family = binomial("probit"),
  lambda,
  corstr = "independence",
  beta.ini = NULL,
  R = NULL,
  scale.fix = TRUE,
  scale.value = 1
)
}
\arguments{
\item{X}{A design matrix of dimension  \code{(nm) * p}.}

\item{y}{A response vector of length  \code{m * n}.}

\item{id}{A vector for identifying subjects/clusters.}

\item{family}{A family object: a list of functions and expressions
for defining link and variance functions. Families supported here is same as
  in \pkg{PGEE} which are binomial, gaussian, gamma and poisson.}

\item{lambda}{A numerical value for the penalization parameter.}

\item{corstr}{A character string, which specifies the type of correlation
structure. Structures supported in PGEE are "AR-1","exchangeable",
"independence", and "unstructured". The default corstr type is "independence".}

\item{beta.ini}{User specified initial values for regression parameters. The default value is NULL.}

\item{R}{User specified correlation matrix. The default value is \code{NULL}.}

\item{scale.fix}{A logical variable; if true, the scale parameter is
fixed at the value of scale.value. The default value is TRUE.}

\item{scale.value}{If  \code{scale.fix = TRUE}, this assignes a
numeric value to which the scale parameter should be fixed.
The default value is 1.}
}
\value{
A list containing the following components:
  \item{betaest}{return final estimation}
  \item{beta_all_step}{return estimate in each iteration}
  \item{inner.count}{iterative count in each stage}
  \item{outer.iter}{iterate number of outer loop}
}
\description{
This function fits a $L_{1}$ penalized GEE model to longitudinal
 data by I-CGD algorithm.
}
\examples{
\dontrun{
set.seed(123)
p <- 256
s <- ceiling(p^{1/3})
n <- ceiling(10 * s * log(p))
m <- 4
# covariance matrix of p number of continuous covariates
X.sigma <- matrix(0, p, p)
{
  for (i in 1:p)
    X.sigma[i,] <- 0.5^(abs((1:p)-i))
}

# generate matrix of covariates
X <- as.matrix(rmvnorm(n*m, mean = rep(0,p), X.sigma))

# true regression parameter associated with the covariate
bt <- runif(s, 0.05, 0.5) # = rep(1/s,s)
beta.true <- c(bt,rep(0,p-s))
# intercept
beta_intercepts <- 0
# unstructure
tt <- runif(m*m,-1,1)
Rtmp <- t(matrix(tt, m,m))\%*\%matrix(tt, m,m)+diag(1,4)
R_tr <- diag(diag(Rtmp)^{-1/2})\%*\%Rtmp\%*\%diag(diag(Rtmp)^{-1/2})
diag(R_tr) = round(diag(R_tr))

# library(SimCorMultRes)
# simulation of clustered binary responses
simulated_binary_dataset <- rbin(clsize = m, intercepts = beta_intercepts,
                                 betas = beta.true, xformula = ~X, cor.matrix = R_tr,
                                 link = "probit")
lambda <- 0.2* s *sqrt(log(p)/n)
data = simulated_binary_dataset$simdata
y = data$y
X = data$X
id = data$id

ptm <- proc.time()
nCGDfit = LassoGEE(X = X, y = y, id = id, family = binomial("probit"),
                 lambda = lambda, corstr = "unstructured")
proc.time() - ptm
betaest <- nCGDfit$betaest
}
}
\references{
Li, Y., Gao, X., and Xu, W. (2020). Statistical consistency for
generalized estimating equation with $L_{1}$ regularization.
}
