% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.LassoGEE.R
\name{print.LassoGEE}
\alias{print.LassoGEE}
\title{print a LassoGEE object}
\usage{
\method{print}{LassoGEE}(x, digits = NULL, ...)
}
\arguments{
\item{x}{fitted 'LassoGEE' object}

\item{digits}{significant digits in printout}

\item{\dots}{additional print arguments}
}
\description{
Print a summary of the results of a LassoGEE model.
}
\details{
A summary of the cross-validated fit is produced. \code{print.cv.LassoGEE(object)}
will print the summary includes Working Correlation and Returned Error Value.
}
\references{
Li, Y., Gao, X., and Xu, W. (2020). Statistical consistency for
generalized estimating equation with \eqn{L_1} regularization.
}
\seealso{
\code{LassoGEE}, and \code{cv.LassoGEE} methods.
}
\keyword{models}
\keyword{regression}
